refactor(admin-ui): move TextAction into its own file
It is a standalone button, not a detail of the actions row - the styling
note about why it is a button rather than the design-system `Link` reads
better next to it than buried above an unrelated component.
