+++ packages/admin-ui/src/FilePicker/primitives/components/previews/index.ts
@@ -2,5 +2,6 @@ export * from "./FilePreview.js";
 export * from "./ImagePreview/index.js";
 export * from "./ItemTextActions.js";
 export * from "./RichItemPreview/index.js";
+export * from "./TextAction.js";
 export * from "./TextOnlyPreview/index.js";
 export * from "./TruncatedFileName.js";
